Preprocessor (#77)
* added List.AppendRange

* added preprocessor (no directives yet though)

* directive errors are now reported

* removed temp file

* fixed test

* removed unused code
