+++ src/Buckle/Compiler.Tests/CodeAnalysis/Parser/ParserTests.cs
@@ -7,7 +7,7 @@ using Xunit;
 namespace Buckle.Tests.CodeAnalysis.Syntax.InternalSyntax;
 
 /// <summary>
-/// Tests on the <see cref="Buckle.CodeAnalysis.Syntax.InternalSyntax.Parser" /> class.
+/// Tests on the <see cref="Buckle.CodeAnalysis.Syntax.InternalSyntax.LanguageParser" /> class.
 /// </summary>
 public sealed class ParserTests {
     [Theory]
