+++ src/Buckle/Compiler/CodeAnalysis/Binding/BoundTree/BoundLabelStatement.cs
@@ -2,7 +2,7 @@
 namespace Buckle.CodeAnalysis.Binding;
 
 /// <summary>
-/// A bound label statement, produced by <see cref="Lowering.Lowerer" />. No <see cref="Syntax.InternalSyntax.Parser" />
+/// A bound label statement, produced by <see cref="Lowering.Lowerer" />. No <see cref="Syntax.InternalSyntax.LanguageParser" />
 /// equivalent.<br/>
 /// E.g.
 /// <code>
