+++ src/Buckle/Compiler/CodeAnalysis/Binding/BoundTree/BoundStatement.cs
@@ -2,7 +2,7 @@
 namespace Buckle.CodeAnalysis.Binding;
 
 /// Note: All bound versions of the <see cref="Syntax.StatementSyntax" /> and <see cref="Syntax.ExpressionSyntax" />
-/// share function with <see cref="Syntax.InternalSyntax.Parser" /> equivalents. Thus use their xml comments for
+/// share function with <see cref="Syntax.InternalSyntax.LanguageParser" /> equivalents. Thus use their xml comments for
 /// reference.
 
 /// <summary>
