+++ src/Buckle/Compiler/CodeAnalysis/Parser/LanguageParser.ScanTemplateArgumentListKind.cs
@@ -1,6 +1,6 @@
 namespace Buckle.CodeAnalysis.Syntax.InternalSyntax;
 
-internal sealed partial class Parser {
+internal sealed partial class LanguageParser {
     private enum ScanTemplateArgumentListKind {
         NotTemplateArgumentList,
         PossibleTemplateArgumentList,
