+++ src/Buckle/Compiler/CodeAnalysis/Parser/SyntaxParser.ResetPoint.cs
@@ -1,0 +1,16 @@
+
+namespace Buckle.CodeAnalysis.Syntax.InternalSyntax;
+
+internal partial class SyntaxParser {
+    protected readonly struct ResetPoint {
+        internal readonly int position;
+        internal readonly GreenNode prevTokenTrailingTrivia;
+
+        internal ResetPoint(
+            int position,
+            GreenNode prevTokenTrailingTrivia) {
+            this.position = position;
+            this.prevTokenTrailingTrivia = prevTokenTrailingTrivia;
+        }
+    }
+}
