+++ src/Buckle/Compiler/CodeAnalysis/Syntax/InternalSyntax/DirectiveTriviaSyntax.cs
@@ -1,0 +1,8 @@
+
+namespace Buckle.CodeAnalysis.Syntax.InternalSyntax;
+
+internal partial class DirectiveTriviaSyntax {
+    internal override DirectiveStack ApplyDirectives(DirectiveStack stack) {
+        return stack.Add(new Directive(this));
+    }
+}
