+++ src/Buckle/Compiler/CodeAnalysis/Syntax/InternalSyntax/SyntaxListBuilder`.cs
@@ -19,6 +19,8 @@ internal sealed class SyntaxListBuilder<T> where T : GreenNode {
         _builder = builder;
     }
 
+    internal SyntaxListBuilder() { }
+
     /// <summary>
     /// Creates a <see cref="SyntaxListBuilder<T>" /> with the default starting size of 8.
     /// </summary>
