+++ src/Buckle/Compiler/SourceCodeKind.cs
@@ -1,0 +1,10 @@
+
+namespace Buckle;
+
+/// <summary>
+/// The type of source.
+/// </summary>
+public enum SourceCodeKind : byte {
+    Regular,
+    Script,
+}
