+++ src/Buckle/Repl/BelteRepl.cs
@@ -165,7 +165,7 @@ public sealed partial class BelteRepl : Repl {
     protected override void EvaluateSubmission(string text) {
         // ONLY use this when evaluating previous submissions, where incremental compilation would do nothing
         // Otherwise, this is much slower than the 0 arity overload
-        var syntaxTree = SyntaxTree.Parse(text);
+        var syntaxTree = SyntaxTree.Parse(text, SourceCodeKind.Script);
         EvaluateSubmissionInternal(syntaxTree);
     }
 
@@ -263,7 +263,7 @@ public sealed partial class BelteRepl : Repl {
     }
 
     private void ClearTree() {
-        state.tree = SyntaxTree.Parse("");
+        state.tree = SyntaxTree.Parse("", SourceCodeKind.Script);
         // This should always be empty by now, but just in case there was a race condition
         _changes.Clear();
     }
